removed typecasts

+++ classes/Board.java
@@ -12,7 +12,7 @@ public class Board {
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<4; j++) {
 				piecesOnBoard.add(new Piece(new int[] {(2*j+(i%2)), i}, players[0]));
-				piecesOnBoard.add(new Piece(new int[] {((int) 7) - 2*j+(i%2), ((int) 7)-i}, players[1]));
+				piecesOnBoard.add(new Piece(new int[] {7 - 2*j+(i%2), 7-i}, players[1]));
 			}
 		}
 	}
